Fix reward structure in config file

+++ src/config/rmax/RmaxConfig.java
@@ -3,7 +3,7 @@ package config.rmax;
 import burlap.behavior.singleagent.auxiliary.performance.PerformanceMetric;
 
 public class RmaxConfig {
-    public int vmax;
+    public double vmax;
     public int threshold;
     public double max_delta;
     public double max_delta_in_model;
